WIP Refactor shader code
