+++ polynomial_symmetry/Coefficients.js
@@ -1,7 +1,7 @@
 // A list of (freq_n, freq_m, amp, phase) coefficients
 // representing complex numbers a_nm in a complex polynomial
 // f(z) = sum_nm a_nm z^n conj(z)^m
-class Coefficients {
+export class Coefficients {
     constructor(tuples) {
         this._tuples = tuples;
     }
